changed contact button to a menu, lots of styles and small tweaks, way too much...
