+++ src/app/components/about-page/about-page.component.html
@@ -2,7 +2,6 @@
 <div class="app-page-container about-page">
   <div class="heading-container">
     <h1 class="app-heading">about</h1>
-    <app-contact-button></app-contact-button>
   </div>
   <div class="section-container">
     <section *ngFor="let section of sections; index as i">
