+++ src/app/components/contact-button/contact-button.component.html
@@ -1,9 +1,109 @@
-<!-- <a
-  role="button"
-  href="mailto:jmoody1813@gmail.com"
+<button
+  [class.active]="menuIsOpen"
   class="contact-button"
-  target="_blank"
-  >e-mail me</a
-> -->
-
-<button>contact</button>
+  (click)="toggleMenu()"
+>
+  contact
+</button>
+<div
+  (click)="toggleMenu()"
+  class="contact-menu-wrapper"
+  [class.active]="menuIsOpen"
+>
+  <div
+    class="circular-menu"
+    [class.active]="menuIsOpen"
+    [style.visibility]="menuIsOpen ? 'visible' : 'hidden'"
+    (click)="stopProp($event)"
+  >
+    <menu class="items-wrapper">
+      <a
+        href="https://github.com/MoodyJW"
+        target="_blank"
+        rel="noopener noreferrer"
+        class="menu-item"
+        ><svg
+          xmlns="http://www.w3.org/2000/svg"
+          class="icon icon-tabler icon-tabler-brand-github"
+          width="44"
+          height="44"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          fill="none"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+        >
+          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+          <path
+            d="M9 19c-4.3 1.4 -4.3 -2.5 -6 -3m12 5v-3.5c0 -1 .1 -1.4 -.5 -2c2.8 -.3 5.5 -1.4 5.5 -6a4.6 4.6 0 0 0 -1.3 -3.2a4.2 4.2 0 0 0 -.1 -3.2s-1.1 -.3 -3.5 1.3a12.3 12.3 0 0 0 -6.2 0c-2.4 -1.6 -3.5 -1.3 -3.5 -1.3a4.2 4.2 0 0 0 -.1 3.2a4.6 4.6 0 0 0 -1.3 3.2c0 4.6 2.7 5.7 5.5 6c-.6 .6 -.6 1.2 -.5 2v3.5"
+          /></svg
+      ></a>
+      <a
+        href="https://www.linkedin.com/in/jasonwmoody/"
+        target="_blank"
+        rel="noopener noreferrer"
+        class="menu-item"
+        ><svg
+          xmlns="http://www.w3.org/2000/svg"
+          class="icon icon-tabler icon-tabler-brand-linkedin"
+          width="44"
+          height="44"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          fill="none"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+        >
+          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+          <rect x="4" y="4" width="16" height="16" rx="2" />
+          <line x1="8" y1="11" x2="8" y2="16" />
+          <line x1="8" y1="8" x2="8" y2="8.01" />
+          <line x1="12" y1="16" x2="12" y2="11" />
+          <path d="M16 16v-3a2 2 0 0 0 -4 0" /></svg
+      ></a>
+      <a
+        href="mailto:jmoody1813@gmail.com"
+        target="_blank"
+        rel="noopener noreferrer"
+        class="menu-item"
+        ><svg
+          xmlns="http://www.w3.org/2000/svg"
+          class="icon icon-tabler icon-tabler-mail"
+          width="44"
+          height="44"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          fill="none"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+        >
+          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+          <rect x="3" y="5" width="18" height="14" rx="2" />
+          <polyline points="3 7 12 13 21 7" /></svg
+      ></a>
+      <a
+        href="https://leetcode.com/MoodyJW/"
+        target="_blank"
+        rel="noopener noreferrer"
+        class="menu-item"
+        ><svg
+          xmlns="http://www.w3.org/2000/svg"
+          xmlns:xlink="http://www.w3.org/1999/xlink"
+          width="44"
+          height="44"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          fill="none"
+        >
+          <path
+            d="M13.483 0a1.374 1.374 0 0 0-.961.438L7.116 6.226l-3.854 4.126a5.266 5.266 0 0 0-1.209 2.104a5.35 5.35 0 0 0-.125.513a5.527 5.527 0 0 0 .062 2.362a5.83 5.83 0 0 0 .349 1.017a5.938 5.938 0 0 0 1.271 1.818l4.277 4.193l.039.038c2.248 2.165 5.852 2.133 8.063-.074l2.396-2.392c.54-.54.54-1.414.003-1.955a1.378 1.378 0 0 0-1.951-.003l-2.396 2.392a3.021 3.021 0 0 1-4.205.038l-.02-.019l-4.276-4.193c-.652-.64-.972-1.469-.948-2.263a2.68 2.68 0 0 1 .066-.523a2.545 2.545 0 0 1 .619-1.164L9.13 8.114c1.058-1.134 3.204-1.27 4.43-.278l3.501 2.831c.593.48 1.461.387 1.94-.207a1.384 1.384 0 0 0-.207-1.943l-3.5-2.831c-.8-.647-1.766-1.045-2.774-1.202l2.015-2.158A1.384 1.384 0 0 0 13.483 0zm-2.866 12.815a1.38 1.38 0 0 0-1.38 1.382a1.38 1.38 0 0 0 1.38 1.382H20.79a1.38 1.38 0 0 0 1.38-1.382a1.38 1.38 0 0 0-1.38-1.382z"
+            fill="none"
+          /></svg
+      ></a>
+    </menu>
+  </div>
+</div>
