+++ src/app/components/home-page/home-page.component.html
@@ -14,6 +14,5 @@
         impostor syndrome
       </p>
     </div>
-    <app-contact-button class="contact-button"></app-contact-button>
   </div>
 </div>
