+++ src/app/components/navigation/navigation.component.scss
@@ -10,6 +10,6 @@
 
   .horizontal-line {
     width: 100%;
-    margin: 0 1em;
+    margin: 0.3em 1em 0 1em;
   }
 }
