+++ src/app/components/navigation/navigation.constants.ts
@@ -1,0 +1,23 @@
+export interface NavLink {
+  url: string;
+  title: string;
+}
+
+export const ROUTER_LINKS: NavLink[] = [
+  {
+    url: '/home',
+    title: 'home',
+  },
+  {
+    url: '/about',
+    title: 'about',
+  },
+  {
+    url: '/projects',
+    title: 'projects',
+  },
+  {
+    url: '',
+    title: 'home',
+  },
+];
