+++ src/app/components/projects-page/projects-page.component.html
@@ -2,7 +2,6 @@
 <div class="app-page-container projects-page">
   <div class="heading-container">
     <h1 class="app-heading">projects</h1>
-    <app-contact-button></app-contact-button>
   </div>
   <div class="project-cards">
     <ng-container *ngFor="let project of projects; index as i">
