+++ src/styles.scss
@@ -29,9 +29,9 @@ $breakpoints: (
 
 :root {
   --primary-background-color: #333;
-  --primary-text-color: rgba(255, 255, 255, 0.934);
+  --primary-text-color: hsla(0, 0%, 100%, 0.934);
   --visited-text-color: hsla(307, 53%, 42%, 0.797);
-  --secondary-text-color: rgba(223, 201, 222, 0.934);
+  --secondary-text-color: hsla(303, 26%, 83%, 0.934);
   --primary-button-color: hsl(312, 26%, 56%);
   --shadow-color: hsl(312, 26%, 56%);
 }
@@ -80,6 +80,12 @@ a:active {
   color: inherit;
 }
 
+a:hover {
+  color: var(--shadow-color);
+  text-shadow: 0px 2px 10px var(--primary-background-color);
+  transform: scale(1.1);
+}
+
 p {
   font-family: "Roboto Mono", monospace;
 }
@@ -123,7 +129,11 @@ p {
 }
 
 .glowing {
-  text-shadow: 0px 2px 10px;
+  color: var(--shadow-color);
+  font-weight: 600;
+  text-shadow: 0px 2px 10px var(--primary-background-color);
+  transform: scale(1.1);
+  pointer-events: none;
 }
 
 @include respond-to("medium") {
